Сделал рефакторинг. Переименовал класс, вынес внутренние класс.
