+++ wear/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".NewTimerActivity"/>
+        <activity android:name=".SetupTimerActivity"/>
         <service android:name=".TimerNotificationService"/>
     </application>
